+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.3.8
 
 - Add some macro and build-in function.
+- Add hover information for macro.
 
 ## v0.3.7